refactor(security): unify secret redaction into single module (#2381)
Consolidate redaction logic from runner.ts, debug.ts, and
onboard-session.ts into a single src/lib/redact.ts module. Adding a
new token pattern now means updating secret-patterns.ts only — no more
parallel changes across 4 files.

- Create src/lib/redact.ts with redact() (partial), redactFull(),
  redactSensitiveText(), redactUrl(), redactError(), writeRedactedResult()
- runner.ts: remove 87 lines of inline redaction, require from redact.ts
- debug.ts: remove REDACT_PATTERNS + redact(), import redactFull
- onboard-session.ts: remove redactSensitiveText + redactUrl, import
- shields-audit.ts: import redact from ./redact instead of ./runner
- debug.sh: delegate to compiled redact module when node available,
  slim sed fallback for environments without node
- secret-patterns.ts: update docs, reduce EXPECTED_SHELL_PREFIXES to
  match the fallback set

All existing tests pass unchanged — pure structural refactor.
