+++ src/lib/runner.ts
@@ -217,91 +217,8 @@ function runArrayCapture(cmd, opts = {}) {
   }
 }
 
-/**
- * Redact known secret patterns from a string to prevent accidental leaks
- * in CLI log and error output. Covers NVIDIA API keys, bearer tokens,
- * generic API key assignments, and base64-style long tokens.
- */
-// Single source of truth for secret patterns — see secret-patterns.ts
-const { SECRET_PATTERNS } = require("./secret-patterns");
-
-/**
- * Partially redact a matched secret string: keep the first 4 chars and replace
- * the rest with asterisks (capped at 20 asterisks).
- */
-function redactMatch(match) {
-  return match.slice(0, 4) + "*".repeat(Math.min(match.length - 4, 20));
-}
-
-/**
- * Redact credentials from a URL string: clears url.password and blanks
- * known auth-style query params (auth, sig, signature, token, access_token).
- * Returns the original value unchanged if it cannot be parsed as a URL.
- */
-function redactUrl(value) {
-  if (typeof value !== "string" || value.length === 0) return value;
-  try {
-    const url = new URL(value);
-    if (url.password) {
-      url.password = "****";
-    }
-    for (const key of [...url.searchParams.keys()]) {
-      if (/(^|[-_])(?:signature|sig|token|auth|access_token)$/i.test(key)) {
-        url.searchParams.set(key, "****");
-      }
-    }
-    return url.toString();
-  } catch {
-    return value;
-  }
-}
-
-/**
- * Redact known secret patterns and authenticated URLs from a string.
- * Non-string values are returned unchanged.
- */
-function redact(str) {
-  if (typeof str !== "string") return str;
-  let out = str.replace(/https?:\/\/[^\s'"]+/g, redactUrl);
-  for (const pat of SECRET_PATTERNS) {
-    out = out.replace(pat, redactMatch);
-  }
-  return out;
-}
-
-/**
- * Redact sensitive fields on an error object before surfacing it to callers.
- * NOTE: this mutates the original error instance in place.
- */
-function redactError(err) {
-  if (!err || typeof err !== "object") return err;
-  const originalMessage = typeof err.message === "string" ? err.message : null;
-  if (typeof err.message === "string") err.message = redact(err.message);
-  if (typeof err.cmd === "string") err.cmd = redact(err.cmd);
-  if (typeof err.stdout === "string") err.stdout = redact(err.stdout);
-  if (typeof err.stderr === "string") err.stderr = redact(err.stderr);
-  if (Array.isArray(err.output)) {
-    err.output = err.output.map((value) => (typeof value === "string" ? redact(value) : value));
-  }
-  if (originalMessage && typeof err.stack === "string") {
-    err.stack = err.stack.replaceAll(originalMessage, err.message);
-  }
-  return err;
-}
-
-/**
- * Write redacted stdout/stderr from a spawnSync result to the parent process streams.
- * No-op when stdio is 'inherit' or not an array.
- */
-function writeRedactedResult(result, stdio) {
-  if (!result || stdio === "inherit" || !Array.isArray(stdio)) return;
-  if (stdio[1] === "pipe" && result.stdout) {
-    process.stdout.write(redact(result.stdout.toString()));
-  }
-  if (stdio[2] === "pipe" && result.stderr) {
-    process.stderr.write(redact(result.stderr.toString()));
-  }
-}
+// Unified redaction — see redact.ts (#2381).
+const { redact, redactError, writeRedactedResult } = require("./redact");
 
 /**
  * Shell-quote a value for safe interpolation into bash -c strings.
